refactor: AI 요약 카드 본문 단일 파일로 통합
